Change DataType Funds table some_int_to_long

+++ Application/Common/CreateDtoes/CreateFundRelProjectDto.cs
@@ -6,8 +6,8 @@
         //درصد پیشرفت فیزیکی
         public int PercenPhysicalProgress { get; set; }
         //مبلغ هزینه شده
-        public int AmountSpent { get; set; }
+        public long AmountSpent { get; set; }
         //بر آورد کل مبلغ مورد نیاز
-        public int TotalAmountNeeded { get; set; }
+        public long TotalAmountNeeded { get; set; }
     }
 }
