+++ Application/Project_IncompletedService/Query/GetProject_IncompletedById.cs
@@ -173,9 +173,9 @@ namespace Application.Project_IncompletedService.Query
         //درصد پیشرفت فیزیکی
         public int PercenPhysicalProgress { get; set; }
         //مبلغ هزینه شده
-        public int AmountSpent { get; set; }
+        public long AmountSpent { get; set; }
         //بر آورد کل مبلغ مورد نیاز
-        public int TotalAmountNeeded { get; set; }
+        public long TotalAmountNeeded { get; set; }
     }
 }
