+++ Domain/ProjectNeeds/Fund.cs
@@ -17,9 +17,9 @@ namespace Domain.ProjectNeeds
         //درصد پیشرفت فیزیکی
         public int PercenPhysicalProgress { get; set; }
         //مبلغ هزینه شده
-        public int AmountSpent { get; set; }
+        public long AmountSpent { get; set; }
         //بر آورد کل مبلغ مورد نیاز
-        public int TotalAmountNeeded { get; set; }
+        public long TotalAmountNeeded { get; set; }
         //
         public int ProjectId { get; set; }
         public Project Project { get; set; }
